Split tests to match FormattingEndpoints
